# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Jimce API",
    "version": "0.0.0-dev"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token sent as Authorization: Bearer <token>"
      }
    },
    "schemas": {
      "400-bad-request-response": {
        "description": "A response sent by the server when your request is malformed (Code 400)",
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "const": 400
          },
          "code": {
            "description": "Internal code describing the error",
            "examples": [
              "FST_ERR_VALIDATION"
            ],
            "type": "string"
          },
          "error": {
            "type": "string",
            "const": "Bad Request"
          },
          "message": {
            "description": "Error message describing what went wrong",
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "error",
          "message"
        ],
        "additionalProperties": false
      },
      "401-unauthorized-response": {
        "description": "A response sent by the server when your request is not authorized",
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "const": 401
          },
          "code": {
            "description": "Internal code describing the error",
            "examples": [
              "FST_JWT_NO_AUTHORIZATION_IN_HEADER"
            ],
            "type": "string"
          },
          "error": {
            "type": "string",
            "const": "Unauthorized"
          },
          "message": {
            "description": "Error message describing what went wrong",
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "error",
          "message"
        ],
        "additionalProperties": false
      },
...
      "500-internal-server-error-response": {
        "description": "A response sent by the server when the server failed to process your request",
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "const": 500
          },
          "code": {
            "description": "Internal code describing the error",
            "examples": [
              "FST_ERR_RESPONSE_SERIALIZATION"
            ],
            "type": "string"
          },
          "error": {
            "type": "string",
            "const": "Internal Server Error"
          },
          "message": {
            "description": "Error message describing what went wrong",
            "examples": [
              "Response does not match the schema"
            ],
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "error",
          "message"
        ],
        "additionalProperties": false
      },
      "success-indicator-response": {
        "description": "A response sent by the server when your request is valid to indicate whether it was successful or not",
        "type": "object",
        "properties": {
          "success": {
            "description": "States whether the requested action was successful",
            "type": "boolean"
          },
          "reason": {
            "description": "An optional reason for the success or fail of your requested action",
            "examples": [
              "User already exists"
            ],
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
...
    "/api/auth/change-password": {
      "put": {
        "summary": "Changes the password",
        "description": "Changes the password for the currently logged in user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newPassword": {
                    "description": "The new password as plain text for the user",
                    "examples": [
                      "123456789abc"
                    ],
                    "type": "string",
                    "minLength": 12
                  }
                },
                "required": [
                  "newPassword"
                ]
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success-indicator-response"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/400-bad-request-response"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/401-unauthorized-response"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/500-internal-server-error-response"
                }
              }
            }
          }
        }
      }
    },
...
  }
}